TST: Flatten test directory, place all `requires_epics` tests in live_tests
