+++ run_tests.py
@@ -11,6 +11,13 @@ if __name__ == '__main__':
     if len(sys.argv) > 1:
         args.extend(sys.argv[1:])
 
+    #Ignore live tests unless given the live keyword
+    if '--live' in args:
+        args.remove('--live')
+        args.append('tests_live')
+    else:
+        args.append('--ignore=tests_live')
+
     print('pytest arguments: {}'.format(args))
 
     sys.exit(pytest.main(args))
