+++ tests/conftest.py
@@ -10,19 +10,6 @@ from pcdsdevices import (ImsMotor, GateValve, Slits, Attenuator,
 from pcdsdevices.epics.areadetector.detectors import (FeeOpalDetector,
                                                        PulnixDetector)
 
-try:
-    import epics
-    pv = epics.PV("XCS:USR:MMS:01")
-    try:
-        val = pv.get()
-    except:
-        val = None
-except:
-    val = None
-epics_subnet = val is not None
-requires_epics = pytest.mark.skipif(not epics_subnet,
-                                    reason="Could not connect to sample PV")
-
 class Params:
     registry = OrderedDict()
