+++ tests/test_signal.py
