+++ tests/test_sim_detectors.py
