+++ tests/test_state.py
@@ -6,7 +6,6 @@ from ophyd.signal import Signal
 
 from pcdsdevices.epics import state
 
-from conftest import requires_epics
 
 
 class PrefixSignal(Signal):
@@ -72,11 +71,3 @@ def test_statesrecord_class():
     state.statesrecord_class("Classname", "state0", "state1", "state2")
 
 
-@requires_epics
-@pytest.mark.timeout(3)
-def test_statesrecord_class_reads():
-    """
-    Instantiate one if we can and make sure value makes sense.
-    """
-    inout = state.InOutStates("XCS:SB2:PIM6:DIODE")
-    assert(inout.value in ("IN", "OUT", "Unknown"))
