+++ tests_live/test_fee_opal_detector.py
