+++ tests_live/test_pulnix_detector.py
